feat: add cross-platform SkillsStore to shared library
